Make css-in-js amp only now that carousel is compatible
